Upgrade Invoice service to .NET 6
