+++ src/InvoiceService/CommunicationChannels/IEMailCommunicator.cs
@@ -1,13 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Net.Mail;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace Pitstop.InvoiceService.CommunicationChannels;
 
-namespace Pitstop.InvoiceService.CommunicationChannels
+public interface IEmailCommunicator
 {
-    public interface IEmailCommunicator
-    {
-        Task SendEmailAsync(MailMessage message);
-    }
+    Task SendEmailAsync(MailMessage message);
 }
