+++ src/InvoiceService/Events/CustomerRegistered.cs
@@ -1,24 +1,20 @@
-﻿using Pitstop.Infrastructure.Messaging;
-using System;
+﻿namespace Pitstop.InvoiceService.Events;
 
-namespace Pitstop.InvoiceService.Events
+public class CustomerRegistered : Event
 {
-    public class CustomerRegistered : Event
+    public readonly string CustomerId;
+    public readonly string Name;
+    public readonly string Address;
+    public readonly string PostalCode;
+    public readonly string City;
+
+
+    public CustomerRegistered(Guid messageId, string customerId, string name, string address, string postalCode, string city) : base(messageId)
     {
-        public readonly string CustomerId;
-        public readonly string Name;
-        public readonly string Address;
-        public readonly string PostalCode;
-        public readonly string City;
-
-
-        public CustomerRegistered(Guid messageId, string customerId, string name, string address, string postalCode, string city) : base(messageId)
-        {
-            CustomerId = customerId;
-            Name = name;
-            Address = address;
-            PostalCode = postalCode;
-            City = city;
-        }
+        CustomerId = customerId;
+        Name = name;
+        Address = address;
+        PostalCode = postalCode;
+        City = city;
     }
 }
