+++ src/InvoiceService/Events/MaintenanceJobFinished.cs
@@ -1,20 +1,16 @@
-﻿using Pitstop.Infrastructure.Messaging;
-using System;
+﻿namespace Pitstop.InvoiceService.Events;
 
-namespace Pitstop.InvoiceService.Events
+public class MaintenanceJobFinished : Event
 {
-    public class MaintenanceJobFinished : Event
-    {
-        public readonly string JobId;
-        public readonly DateTime StartTime;
-        public readonly DateTime EndTime;
+    public readonly string JobId;
+    public readonly DateTime StartTime;
+    public readonly DateTime EndTime;
 
-        public MaintenanceJobFinished(Guid messageId, string jobId, DateTime startTime, DateTime endTime) : 
-            base(messageId)
-        {
-            JobId = jobId;
-            StartTime = startTime;
-            EndTime = endTime;
-        }
+    public MaintenanceJobFinished(Guid messageId, string jobId, DateTime startTime, DateTime endTime) :
+        base(messageId)
+    {
+        JobId = jobId;
+        StartTime = startTime;
+        EndTime = endTime;
     }
 }
