+++ src/InvoiceService/GlobalUsings.cs
@@ -1,0 +1,19 @@
+global using Pitstop.InvoiceService;
+global using Pitstop.InvoiceService.Events;
+global using Pitstop.InvoiceService.Model;
+global using Pitstop.Infrastructure.Messaging;
+global using Pitstop.Infrastructure.Messaging.Configuration;
+global using Pitstop.InvoiceService.CommunicationChannels;
+global using Pitstop.InvoiceService.Repositories;
+global using Microsoft.Extensions.Configuration;
+global using Microsoft.Extensions.DependencyInjection;
+global using Microsoft.Extensions.Hosting;
+global using Serilog;
+global using Newtonsoft.Json.Linq;
+global using System.Net;
+global using System.Net.Mail;
+global using System.Net.Mime;
+global using System.Text;
+global using Dapper;
+global using Polly;
+global using System.Data.SqlClient;
