+++ src/InvoiceService/Model/Customer.cs
@@ -1,16 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿namespace Pitstop.InvoiceService.Model;
 
-namespace Pitstop.InvoiceService.Model
+public class Customer
 {
-    public class Customer
-    {
-        public string CustomerId { get; set; }
-        public string Name { get; set; }
-        public string Address { get; set; }
-        public string PostalCode { get; set; }
-        public string City { get; set; }
-    }
+    public string CustomerId { get; set; }
+    public string Name { get; set; }
+    public string Address { get; set; }
+    public string PostalCode { get; set; }
+    public string City { get; set; }
 }
+
