+++ src/InvoiceService/Model/Invoice.cs
@@ -1,16 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace Pitstop.InvoiceService.Model;
 
-namespace Pitstop.InvoiceService.Model
+public class Invoice
 {
-    public class Invoice
-    {
-        public string InvoiceId { get; set; }
-        public DateTime InvoiceDate { get; set; }
-        public string CustomerId { get; set; }
-        public decimal Amount { get; set; }
-        public string Specification { get; set; }
-        public string JobIds { get; set; }
-    }
+    public string InvoiceId { get; set; }
+    public DateTime InvoiceDate { get; set; }
+    public string CustomerId { get; set; }
+    public decimal Amount { get; set; }
+    public string Specification { get; set; }
+    public string JobIds { get; set; }
 }
