+++ src/InvoiceService/Program.cs
@@ -1,58 +1,32 @@
-﻿using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Pitstop.Infrastructure.Messaging;
-using Pitstop.Infrastructure.Messaging.Configuration;
-using Pitstop.InvoiceService.CommunicationChannels;
-using Pitstop.InvoiceService.Repositories;
-using Serilog;
-using System;
-using System.IO;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Pitstop.InvoiceService
-{
-    class Program
+﻿IHost host = Host
+    .CreateDefaultBuilder(args)
+    .ConfigureServices((hostContext, services) =>
     {
-        public static async Task Main(string[] args)
+        services.UseRabbitMQMessageHandler(hostContext.Configuration);
+
+        services.AddTransient<IInvoiceRepository>((svc) =>
         {
-            var host = CreateHostBuilder(args).Build();
-            await host.RunAsync();
-        }
+            var sqlConnectionString = hostContext.Configuration.GetConnectionString("InvoiceServiceCN");
+            return new SqlServerInvoiceRepository(sqlConnectionString);
+        });
 
-        private static IHostBuilder CreateHostBuilder(string[] args)
+        services.AddTransient<IEmailCommunicator>((svc) =>
         {
-            var hostBuilder = Host.CreateDefaultBuilder(args)
-                .ConfigureServices((hostContext, services) =>
-                {
-                    services.UseRabbitMQMessageHandler(hostContext.Configuration);
+            var mailConfigSection = hostContext.Configuration.GetSection("Email");
+            string mailHost = mailConfigSection["Host"];
+            int mailPort = Convert.ToInt32(mailConfigSection["Port"]);
+            string mailUserName = mailConfigSection["User"];
+            string mailPassword = mailConfigSection["Pwd"];
+            return new SMTPEmailCommunicator(mailHost, mailPort, mailUserName, mailPassword);
+        });
 
-                    services.AddTransient<IInvoiceRepository>((svc) =>
-                    {
-                        var sqlConnectionString = hostContext.Configuration.GetConnectionString("InvoiceServiceCN");
-                        return new SqlServerInvoiceRepository(sqlConnectionString);
-                    });
+        services.AddHostedService<InvoiceWorker>();
+    })
+    .UseSerilog((hostContext, loggerConfiguration) =>
+    {
+        loggerConfiguration.ReadFrom.Configuration(hostContext.Configuration);
+    })
+    .UseConsoleLifetime()
+    .Build();
 
-                    services.AddTransient<IEmailCommunicator>((svc) =>
-                    {
-                        var mailConfigSection = hostContext.Configuration.GetSection("Email");
-                        string mailHost = mailConfigSection["Host"];
-                        int mailPort = Convert.ToInt32(mailConfigSection["Port"]);
-                        string mailUserName = mailConfigSection["User"];
-                        string mailPassword = mailConfigSection["Pwd"];
-                        return new SMTPEmailCommunicator(mailHost, mailPort, mailUserName, mailPassword);
-                    });
-
-                    services.AddHostedService<InvoiceManager>();
-                })
-                .UseSerilog((hostContext, loggerConfiguration) =>
-                {
-                    loggerConfiguration.ReadFrom.Configuration(hostContext.Configuration);
-                })
-                .UseConsoleLifetime();
-
-            return hostBuilder;
-        }
-    }
-}
+await host.RunAsync();
