+++ src/InvoiceService/Repositories/IInvoiceRepository.cs
@@ -1,18 +1,11 @@
-﻿using Pitstop.InvoiceService.Model;
-using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using System.Transactions;
+﻿namespace Pitstop.InvoiceService.Repositories;
 
-namespace Pitstop.InvoiceService.Repositories
+public interface IInvoiceRepository
 {
-    public interface IInvoiceRepository
-    {
-        Task RegisterCustomerAsync(Customer customer);
-        Task<Customer> GetCustomerAsync(string customerId);
-        Task RegisterMaintenanceJobAsync(MaintenanceJob job);
-        Task MarkMaintenanceJobAsFinished(string jobId, DateTime startTime, DateTime endTime);
-        Task<IEnumerable<MaintenanceJob>> GetMaintenanceJobsToBeInvoicedAsync();
-        Task RegisterInvoiceAsync(Invoice invoice);
-    }
+    Task RegisterCustomerAsync(Customer customer);
+    Task<Customer> GetCustomerAsync(string customerId);
+    Task RegisterMaintenanceJobAsync(MaintenanceJob job);
+    Task MarkMaintenanceJobAsFinished(string jobId, DateTime startTime, DateTime endTime);
+    Task<IEnumerable<MaintenanceJob>> GetMaintenanceJobsToBeInvoicedAsync();
+    Task RegisterInvoiceAsync(Invoice invoice);
 }
