+++ src/InvoiceService/Repositories/SqlServerInvoiceRepository.cs
@@ -1,15 +1,4 @@
-﻿using System;
-using System.Threading.Tasks;
-using Dapper;
-using Pitstop.InvoiceService.Model;
-using Polly;
-using System.Data.SqlClient;
-using System.Collections.Generic;
-using System.Transactions;
-using System.Linq;
-using Serilog;
-
-namespace Pitstop.InvoiceService.Repositories
+﻿namespace Pitstop.InvoiceService.Repositories
 {
     public class SqlServerInvoiceRepository : IInvoiceRepository
     {
